Add a switch to disable PSPs via values
We require PSPs to be installed on k8s < 1.25, and since in the flux-app
1.2.0 PSPs were dumped completely, it can't be installed on most of our
clusters anymore. Adding PSPs to the 1.2.* seems like a bad idea, since
they were removed because of deprecation. That's why I'm adding a switch
as a patch to 1.1.*. During the migration we will be able to use this
version, and once 1.25 is our default k8s version, we can upgrade to
flux-app 1.2.* or greater